Generate package endpoint
Instead of dynamically serving the package endpoint, an `index.json` file is pregenerated when run `mage build` for each package. This simplifies the serving code.

A package `p` was created to share the code. The naming is not optimal but `package` cannot be used as it is a reserved term. A better structure will be found later.

The `.tar.gz` handler can also be removed, as it was only there as a placeholder for the matching of the endpoint. Now that `/package/{name}` endpoint is removed, it's not needed anymore. Overall there are now only 2 endpoints left. CatchAll to match files and the `/search` endpoint handler.
